+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Upgrade the `org.realityforge.org.jetbrains.annotations` artifact to version `1.6.0`.
+* Handle the return of `null` from `toString()` in `StringifyReplacer`.
 
 ### [v0.118](https://github.com/arez/arez-spytools/tree/v0.118) (2020-08-14) · [Full Changelog](https://github.com/arez/arez-spytools/compare/v0.117...v0.118)